DL-60: Refactor basic logic into basic crud/search services
