+++ src/VirtoCommerce.PricingModule.Data/Services/Crud/PricelistAssignmentService.cs
@@ -1,53 +1,25 @@
 using System;
 using System.Collections.Generic;
-using System.Threading.Tasks;
 using FluentValidation;
-using VirtoCommerce.Platform.Caching;
 using VirtoCommerce.Platform.Core.Caching;
-using VirtoCommerce.Platform.Core.Common;
 using VirtoCommerce.Platform.Core.Events;
-using VirtoCommerce.Platform.Data.GenericCrud;
 using VirtoCommerce.PricingModule.Core.Events;
 using VirtoCommerce.PricingModule.Core.Model;
 using VirtoCommerce.PricingModule.Data.Model;
 using VirtoCommerce.PricingModule.Data.Repositories;
+using VirtoCommerce.PricingModule.Data.Services.Crud.Basic;
 
 namespace VirtoCommerce.PricingModule.Data.Services
 {
-    public class PricelistAssignmentService : CrudService<PricelistAssignment, PricelistAssignmentEntity, PricelistAssignmentChangingEvent, PricelistAssignmentChangedEvent>
+    public class PricelistAssignmentService : BasicPricelistAssignmentService<PricelistAssignment, PricelistAssignmentEntity, PricelistAssignmentChangingEvent, PricelistAssignmentChangedEvent>
     {
-        private readonly AbstractValidator<IEnumerable<PricelistAssignment>> _validator;
 
         public PricelistAssignmentService(Func<IPricingRepository> repositoryFactory,
             IPlatformMemoryCache platformMemoryCache,
             IEventPublisher eventPublisher,
             AbstractValidator<IEnumerable<PricelistAssignment>> validator)
-            : base(repositoryFactory, platformMemoryCache, eventPublisher)
+            : base(repositoryFactory, platformMemoryCache, eventPublisher, validator)
         {
-            _validator = validator;
-        }
-
-        protected override async Task<IEnumerable<PricelistAssignmentEntity>> LoadEntities(IRepository repository, IEnumerable<string> ids, string responseGroup)
-        {
-            return await ((IPricingRepository)repository).GetPricelistAssignmentsByIdAsync(ids);
-        }
-
-        protected override Task BeforeSaveChanges(IEnumerable<PricelistAssignment> models)
-        {
-            _validator.ValidateAndThrow(models);
-
-            return base.BeforeSaveChanges(models);
-        }
-
-        protected override void ClearCache(IEnumerable<PricelistAssignment> models)
-        {
-            foreach (var assignment in models)
-            {
-                GenericCachingRegion<PricelistAssignment>.ExpireTokenForKey(assignment.Id);
-                GenericCachingRegion<Pricelist>.ExpireTokenForKey(assignment.PricelistId);
-            }
-            GenericSearchCachingRegion<PricelistAssignment>.ExpireRegion();
-            GenericCachingRegion<PricelistAssignment>.ExpireRegion();
         }
     }
 }
