+++ src/VirtoCommerce.PricingModule.Data/Services/Crud/PricelistService.cs
@@ -1,27 +1,19 @@
 using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 using VirtoCommerce.Platform.Core.Caching;
-using VirtoCommerce.Platform.Core.Common;
 using VirtoCommerce.Platform.Core.Events;
-using VirtoCommerce.Platform.Data.GenericCrud;
 using VirtoCommerce.PricingModule.Core.Events;
 using VirtoCommerce.PricingModule.Core.Model;
 using VirtoCommerce.PricingModule.Data.Model;
 using VirtoCommerce.PricingModule.Data.Repositories;
+using VirtoCommerce.PricingModule.Data.Services.Crud.Basic;
 
 namespace VirtoCommerce.PricingModule.Data.Services
 {
-    public class PricelistService : CrudService<Pricelist, PricelistEntity, PricelistChangingEvent, PricelistChangedEvent>
+    public class PricelistService : BasicPricelistService<Pricelist, PricelistEntity, PricelistChangingEvent, PricelistChangedEvent>
     {
         public PricelistService(Func<IPricingRepository> repositoryFactory, IPlatformMemoryCache platformMemoryCache, IEventPublisher eventPublisher)
             : base(repositoryFactory, platformMemoryCache, eventPublisher)
         {
-        }
-
-        protected override async Task<IEnumerable<PricelistEntity>> LoadEntities(IRepository repository, IEnumerable<string> ids, string responseGroup)
-        {
-            return await ((IPricingRepository)repository).GetPricelistByIdsAsync(ids, responseGroup);
         }
     }
 }
