+++ src/VirtoCommerce.PricingModule.Data/Services/Search/PricelistAssignmentSearchService.cs
@@ -9,59 +9,16 @@ using VirtoCommerce.PricingModule.Core.Model;
 using VirtoCommerce.PricingModule.Core.Model.Search;
 using VirtoCommerce.PricingModule.Data.Model;
 using VirtoCommerce.PricingModule.Data.Repositories;
+using VirtoCommerce.PricingModule.Data.Services.Search.Basic;
 
 namespace VirtoCommerce.PricingModule.Data.Services
 {
-    public class PricelistAssignmentSearchService : SearchService<PricelistAssignmentsSearchCriteria, PricelistAssignmentSearchResult, PricelistAssignment, PricelistAssignmentEntity>
+    public class PricelistAssignmentSearchService : BasicPricelistAssignmentSearchService<PricelistAssignmentSearchResult, PricelistAssignment, PricelistAssignmentEntity>
     {
-
         public PricelistAssignmentSearchService(Func<IPricingRepository> repositoryFactory, IPlatformMemoryCache platformMemoryCache,
             ICrudService<PricelistAssignment> pricelistAssignmentService)
            : base(repositoryFactory, platformMemoryCache, pricelistAssignmentService)
         {
-        }
-
-        protected override IQueryable<PricelistAssignmentEntity> BuildQuery(IRepository repository, PricelistAssignmentsSearchCriteria criteria)
-        {
-            var query = ((IPricingRepository)repository).PricelistAssignments;
-
-            if (!criteria.PriceListIds.IsNullOrEmpty())
-            {
-                query = query.Where(x => criteria.PriceListIds.Contains(x.PricelistId));
-            }
-
-            if (!string.IsNullOrEmpty(criteria.Keyword))
-            {
-                query = query.Where(x => x.Name.Contains(criteria.Keyword) || x.Description.Contains(criteria.Keyword));
-            }
-
-            if (!criteria.StoreIds.IsNullOrEmpty())
-            {
-                query = query.Where(x => criteria.StoreIds.Contains(x.StoreId));
-            }
-
-            if (!criteria.CatalogIds.IsNullOrEmpty())
-            {
-                query = query.Where(x => criteria.CatalogIds.Contains(x.CatalogId));
-            }
-
-            return query;
-        }
-
-        protected override IList<SortInfo> BuildSortExpression(PricelistAssignmentsSearchCriteria criteria)
-        {
-            var sortInfos = criteria.SortInfos;
-            if (sortInfos.IsNullOrEmpty())
-            {
-                sortInfos = new[]
-                {
-                    new SortInfo
-                    {
-                        SortColumn = nameof(PricelistAssignment.Priority)
-                    }
-                };
-            }
-            return sortInfos;
         }
     }
 }
